@@ -4,7 +4,7 @@
   "description": "",
   "main": "server.js",
   "engines": {
-    "node": "14.15.3"
+    "node": "18.12.1"
   },
   "scripts": {
     "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
@@ -13,28 +13,18 @@
     "client": "cd client && npm run start",
     "seed": "node scripts/seedDB.js",
     "install": "cd client && yarn install",
-    "build": "cd client && npm run build",
-    "heroku-postbuild": "npm run build"
+    "build": "cd client && npm run build"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
     "concurrently": "^3.5.0",
-    "nodemon": "^1.11.0"
+    "react-scripts": "^5.0.1"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "body-parser": "^1.17.2",
-    "bootstrap": "^4.4.1",
-    "event-emitter": "^0.3.5",
-    "express": "^4.15.4",
-    "firebase": "^7.2.2",
-    "if-env": "^1.0.4",
-    "mongoose": "^5.9.19",
-    "react-modal": "^3.11.1",
-    "react-router-dom": "^5.1.2",
-    "react-toast-notifications": "^2.2.5",
-    "react-toastify": "^5.4.0",
-    "styled-components": "^4.4.0"
+    "axios": "^1.2.3",
+    "express": "^4.18.2",
+    "mongoose": "^6.8.4",
+    "react": "^18.2.0"
   }
 }
