@@ -1,5 +1,5 @@
-import firebase from "firebase/app";
-import "firebase/auth";
+import firebase from "firebase/compat/app";
+import "firebase/compat/auth";
 
 const config = {
   apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
